Organizei o código
Tá lindo boy
