+++ lc.py
@@ -1,4 +1,5 @@
 import pygame as pg
+from time import time
 #Decidi que a escala vai ser 64p:1m
 
 class Wall(pg.sprite.Sprite):
@@ -8,7 +9,19 @@ class Wall(pg.sprite.Sprite):
 		self.image = pg.Surface((abs(self.gs*(pos[0] - pos2[0])), abs(self.gs*(pos[1] - pos2[1]))))
 		self.image.fill((30,30,30))
 		self.rect = self.image.get_rect(x = self.gs*pos[0] , y = -self.gs*pos[1]) #na construção o y tem que ser negativo para facilitar na planta
-	
+
+class Door(pg.sprite.Sprite):
+	def __init__(self, x, y, width, height):
+		super().__init__()
+		self.image = pg.Surface((width, height))
+		self.rect = self.image.get_rect(x = x, y = y)
+		self.open_time = 0.5
+		self.open_delta = 0
+	def open(self):
+		if time() - self.open_delta > self.open_time:
+			self.open_delta = time()
+			self.image = pg.transform.rotate(self.image, 90)
+			self.rect = self.image.get_rect(x = self.rect.x, y = self.rect.y)
 
 
 wall0 = Wall((0, 2), (-1, 1))
